+++ packages/brick_sqlite_abstract/lib/src/sqlite.dart
@@ -1,8 +1,19 @@
 import 'package:brick_core/field_serializable.dart';
 
+import 'package:brick_sqlite_abstract/src/db/migration.dart' show Column;
+export 'package:brick_sqlite_abstract/src/db/migration.dart' show Column;
+
 /// An annotation used to specify how a field is serialized.
 /// Heavily inspired by [JsonKey](https://github.com/dart-lang/json_serializable/blob/master/json_annotation/lib/src/json_key.dart)
 class Sqlite implements FieldSerializable {
+  /// In very exceptional circumstance, the SQLite column type can be declared.
+  ///
+  /// Leaving this field `null` will allow Brick to infer the column type from the
+  /// Type argument. This will not create foreign keys or associations.
+  ///
+  /// Advanced use only.
+  final Column columnType;
+
   @override
   final String defaultValue;
 
@@ -45,6 +56,7 @@ class Sqlite implements FieldSerializable {
   ///
   /// Only required when the default behavior is not desired.
   const Sqlite({
+    this.columnType,
     this.defaultValue,
     this.fromGenerator,
     this.ignore,