+++ packages/brick_sqlite_generators/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Unreleased
 
 * Type arguments are stripped from fields when building the `fieldsToSqliteColumns` definition (#31)
+* Support `columnType` from SQLite annotations